code/requirements.txt
tensorflow-gpu
numpy
opencv-python
pandas
scipy
tqdm
matplotlib
pillow
pyqt5
sklearn
scikit-image
cmake
jupyter
